[tools] Move manual release mirroring to CI
Simplify the script to push the release artifacts to S3 by following
a similar workflow as the `tools/workspace/mirror_to_s3`. Check only
the most recent release when mirroring artifacts.

Promote the script out of `dev/`.
