+++ .github/ISSUE_TEMPLATE/POST_RELEASE_ACTIONS.yml
@@ -1,8 +1,8 @@
 name: Post-Release Actions
 title: Post-release actions for release <version>
 description:
-  Create a tracking issue for Docker image tagging, S3 upload, and Apt
-  repository updates for a new Drake release.
+  Create a tracking issue for Docker image tagging and Apt repository updates
+  for a new Drake release.
 labels: ["type: release"]
 assignees: BetsyMcPhail
 body:
@@ -35,8 +35,6 @@ body:
         completion."
       options:
         - label: "Run push_docker.py script to push Docker images"
-          required: false
-        - label: "Run push_release.py script to mirror the artifacts to S3"
           required: false
         - label: "Run push_release script for Apt repository updates"
           required: false
