+++ tools/release_engineering/dev/BUILD.bazel
@@ -6,16 +6,6 @@ load(
 
 package(default_visibility = ["//visibility:private"])
 
-# This program is used by a small subset of Drake maintainers (and depends on
-# maintainer-only packages) and relies on manual testing.
-drake_py_binary(
-    name = "push_release",
-    srcs = ["push_release.py"],
-    deps = [
-        "@github3_py_internal//:github3_py",
-    ],
-)
-
 # This program is deprecated. Docker images will be removed in the future.
 drake_py_binary(
     name = "push_docker",
